refactor: 컴포넌트 네이밍 수정
전체 물품을 Product로 네이밍한다.
